refactor(web): one component per file via per-component folders
Split Sidebar, RequestForm, and ExampleAccordion (each held >1 component)
into folders named after the component, with the entry file plus siblings
for subcomponents and shared constants. Document the convention in AGENTS.md.
