+++ web/src/components/RequestForm/constants.ts
@@ -1,0 +1,4 @@
+/** Shared styling for the request field tables. */
+export const labelClass = "mt-2 block text-[11px] uppercase tracking-wide text-muted-foreground";
+export const cellInputClass =
+  "h-8 w-full min-w-0 bg-transparent px-2 font-mono text-xs outline-none placeholder:text-muted-foreground/50 focus:bg-muted/60";
