+++ web/src/components/__tests__/RequestForm.test.tsx
@@ -2,7 +2,7 @@
 import { useState } from "react";
 import { describe, it, expect, vi, afterEach } from "vitest";
 import { render, screen, fireEvent, cleanup, within } from "@testing-library/react";
-import { RequestForm } from "../RequestForm";
+import { RequestForm } from "../RequestForm/RequestForm";
 import type { RequestFormState } from "../../request/types";
 
 const form = (over: Partial<RequestFormState> = {}): RequestFormState => ({
